move nvme structs into nvme.h
